requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
pydantic==1.10.11
sqlalchemy==1.4.48
psycopg[binary]
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic==1.10.2
sendgrid==6.11.0
redis==4.5.1
python-multipart==0.0.6
itsdangerous==2.1.2
psycopg2-2.9.11
pip install pydantic[email]
sqlalchemy-2.0.44
greenlet-3.2.4
bcrypt-5.0.0
certifi-2025.11.12
urllib3-2.5.0